requirements.txt
fastapi
uvicorn[standard]
aiogram
openai
python-dotenv
sqlalchemy
aiosqlite
serpapi
httpx
greenlet
pydantic